+++ mcomix/thumbnail_tools.py
@@ -8,25 +8,24 @@ import shutil
 import tempfile
 import mimetypes
 import threading
-import itertools
-import traceback
 import locale
 import PIL.Image as Image
 from urllib.request import pathname2url
-
+from typing import Optional, TYPE_CHECKING
 from hashlib import md5
 
 from mcomix.preferences import prefs
-from mcomix import archive_extractor
 from mcomix import constants
 from mcomix import archive_tools
 from mcomix import tools
 from mcomix import image_tools
 from mcomix import portability
-from mcomix import i18n
 from mcomix import callback
 from mcomix import log
 from mcomix.i18n import _
+
+if TYPE_CHECKING:
+    from gi.repository import GdkPixbuf
 
 
 class Thumbnailer(object):
@@ -69,7 +68,7 @@ class Thumbnailer(object):
         self.force_recreation = force_recreation
         self.archive_support = archive_support
 
-    def thumbnail(self, filepath, threaded=False):
+    def thumbnail(self, filepath: str, threaded: bool = False) -> Optional["GdkPixbuf.Pixbuf"]:
         """ Returns a thumbnail pixbuf for <filepath>, transparently handling
         both normal image files and archives. If a thumbnail file already exists,
         it is re-used. Otherwise, a new thumbnail is created from <filepath>.